Move old proposals to old folder
